+++ tests/unit/iterators/test_iterators.py
@@ -29,5 +29,5 @@ def test_segmentation_iterator(images_v04: dict[str, Path], zarr_name: str):
     iterator = SegmentationIterator(image, label, channel_selection=0, axes_order="yx")
     iterator = iterator.by_yx()
     for i, (img_chunk, writer) in enumerate(iterator.iter_as_numpy()):
-        label_patch = np.full(shape=img_chunk.shape, fill_value=i + 1, dtype=np.uint32)
+        label_patch = np.full(shape=img_chunk.shape, fill_value=i + 1, dtype=np.uint8)
         writer(label_patch)
